pyproject.toml
[tool.poetry]
name = "covid-w-polsce"
version = "0.1.0"
description = ""
authors = ["FlyinButter1 <85359786+FlyinButter1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
streamlit = "^1.30.0"
lxml = "^5.1.0"
azure-functions = "^1.18.0"
azure-storage-blob = "^12.19.1"
azure-identity = "^1.15.0"
fastapi = "^0.110.1"
html5lib = "^1.1"
beautifulsoup4 = "^4.12.3"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"
black = "^23.12.1"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
